refactor(telegram): dedupe prompt media paths
